refactor: domain 패키지로 이동 #160
