+++ src/main/java/com/muud/playlist/domain/PlayList.java
@@ -1,8 +1,8 @@
-package com.muud.playlist.entity;
+package com.muud.playlist.domain;
 
 import com.muud.emotion.domain.Emotion;
 import com.muud.global.common.BaseEntity;
-import com.muud.playlist.dto.VideoDto;
+import com.muud.playlist.domain.dto.VideoDto;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
