+++ src/main/java/com/muud/playlist/domain/dto/VideoDto.java
@@ -1,4 +1,4 @@
-package com.muud.playlist.dto;
+package com.muud.playlist.domain.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import jakarta.persistence.Column;
